neo-vault/services/security-auditor/requirements.txt
fastapi
uvicorn
confluent-kafka
pydantic
scikit-learn
numpy
pandas
sqlalchemy
strawberry-graphql
httpx
python-dotenv
